service/Services/entities/src/main/java/org/infy/idp/entities/jobs/buildinfo/ArtifactRepo.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.idp.entities.jobs.buildinfo;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store artifact repository details for build
 * 
 * @author Infosys
 *
 */
public class ArtifactRepo {
	@SerializedName("repoURL")
	@Expose
	private String repoURL;
	@SerializedName("repoName")
	@Expose
	private String repoName;
	@SerializedName("repoUsername")
	@Expose
	private String repoUsername;
	@SerializedName("repoPassword")
	@Expose
	private String repoPassword;

	public String getRepoURL() {
		return repoURL;
	}

	public void setRepoURL(String repoURL) {
		this.repoURL = repoURL;
	}

	public String getRepoName() {
		return repoName;
	}

	public void setRepoName(String repoName) {
		this.repoName = repoName;
	}

	public String getRepoUsername() {
		return repoUsername;
	}

	public void setRepoUsername(String repoUsername) {
		this.repoUsername = repoUsername;
	}

	public String getRepoPassword() {
		return repoPassword;
	}

	public void setRepoPassword(String repoPassword) {
		this.repoPassword = repoPassword;
	}

}
